backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.7.2
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
jinja2==3.1.4
google-cloud-firestore==2.16.0
google-auth==2.29.0

